+++ crypto/mod.ts
@@ -4,5 +4,5 @@ export * as ss58 from "./ss58.ts"
 // moderate --exclude hex.ts ss58.ts
 
 export * from "./hashers.ts"
-export * from "./Keypair.ts"
+export * from "./Sr25519.ts"
 export * from "./test_pairs.ts"
